backend/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.1
python-dotenv==1.1.0
requests==2.31.0
Werkzeug==3.1.3
